Added new stream operations Cipher_Encrypt_Stream() and Cipher_Decrypt_Stream().
These new functions allow the use of the Cipher_t on file descriptors,
allowing easy access to streaming encryption capabilities. This can be
used for either files, network sockets, or anything capable of being
operated on by the standard read() and write() calls.

Updated Makefiles and build process to derive dependencies from headers,
to ensure that changes to headers trigger the correct rebuilding from
source files.

Fixed some previously hidden bugs in the Cipher_Encrypt() and
Cipher_Decrypt() functions where they did not satisfy the
const-correctness of the In pointer. This lead to corruption when
performing encryption and decryption where the in and out pointers
actually point to the same memory. Also updated the Cipher Tests to
enforce that in-place operation correctly returns the expected result.